Make agent independent processes
