+++ src/alto/common/logging.py
@@ -1,0 +1,6 @@
+import logging
+import sys
+
+def fail_with_msg(level, msg):
+    logging.log(level, msg)
+    sys.exit(-1)
